fixed LNK4042 Warning
