--- tic-tac-toe/x64/Debug/tic-tac-toe.Build.CppClean.log
@@ -1,18 +1,0 @@
-e:\programming projects\c++\tic-tac-toe\tic-tac-toe\x64\debug\vc143.pdb
-e:\programming projects\c++\tic-tac-toe\tic-tac-toe\x64\debug\vc143.idb
-e:\programming projects\c++\tic-tac-toe\tic-tac-toe\x64\debug\main.obj
-e:\programming projects\c++\tic-tac-toe\tic-tac-toe\x64\debug\human_player.obj
-e:\programming projects\c++\tic-tac-toe\tic-tac-toe\x64\debug\grid.obj
-e:\programming projects\c++\tic-tac-toe\tic-tac-toe\x64\debug\comp_player.obj
-e:\programming projects\c++\tic-tac-toe\tic-tac-toe\x64\debug\common.obj
-e:\programming projects\c++\tic-tac-toe\x64\debug\tic-tac-toe.exe
-e:\programming projects\c++\tic-tac-toe\x64\debug\tic-tac-toe.pdb
-e:\programming projects\c++\tic-tac-toe\tic-tac-toe\x64\debug\tic-tac-toe.ilk
-e:\programming projects\c++\tic-tac-toe\tic-tac-toe\x64\debug\tic-tac-toe.tlog\cl.command.1.tlog
-e:\programming projects\c++\tic-tac-toe\tic-tac-toe\x64\debug\tic-tac-toe.tlog\cl.items.tlog
-e:\programming projects\c++\tic-tac-toe\tic-tac-toe\x64\debug\tic-tac-toe.tlog\cl.read.1.tlog
-e:\programming projects\c++\tic-tac-toe\tic-tac-toe\x64\debug\tic-tac-toe.tlog\cl.write.1.tlog
-e:\programming projects\c++\tic-tac-toe\tic-tac-toe\x64\debug\tic-tac-toe.tlog\link.command.1.tlog
-e:\programming projects\c++\tic-tac-toe\tic-tac-toe\x64\debug\tic-tac-toe.tlog\link.read.1.tlog
-e:\programming projects\c++\tic-tac-toe\tic-tac-toe\x64\debug\tic-tac-toe.tlog\link.secondary.1.tlog
-e:\programming projects\c++\tic-tac-toe\tic-tac-toe\x64\debug\tic-tac-toe.tlog\link.write.1.tlog
